Add headers to CMake executable
This is not necessary to build the target. The idea was to maybe help
our IDEs or other tools better understand the dependencies (specifically
CLion), but that idea may be bogus. We should look up whether there is
ever any reason to list header files in CMake's add_executable.

So, in summary this change may be pointless, but it doesn't hurt.