Add Model Viewer program
So far it shows a textured cube in the center of the window. We
plan to add the ability to load arbitrary textured models and use
keyboard input to move and rotate the displayed model. And later
we'll experiment with other effects, like lighting etc.

Much help was drawn from www.learnopengl.com tutorials, and we use
a couple of the utility classes from there.