Start adding code to process Assimp scene, closely following De Vries
Our goal is to start with a simplified version of De Vries' model
class, which holds all the mesh, texture, etc. content that Assimp
extracts from the file it loads. We will try to just implement the
basic features to display the model meshes, and then we will add
more advanced features and details later.