Add basic window setup and render loop to function grapher main
Next: We will add code to build the mesh for displaying the x,y-plane
and the graph of the function z = f(x,y) above it.

To start, we plan to just use a uniform triangular tessellation based
on a rectangular grid, but later we might add the ability to refine
the grid addaptively where the function's output is changing rapidly.