Add initial implementation of function grapher
Displays triangular tessellation of [0,1] x [0,1] in the x,y-plane
(which is the x,z-plane in OpenGL coords), and a mesh interpolation
of the graph of z = f(x,y), for a user-supplied function f.

TODO: We still need to work out some smalle kinks in the graphing
machinery. And, there are things we can do to improve the appearance
of our graph. But, so far so good.